Fix: --debug with LLVM 3.9 (subprogram definitions must have a compile unit)
In LLVM 3.9 DISubprogram references the DICompileUnit directly, we
thus need to generate the DICompileUnit before we generate LLVM IR
code for the rest of the module (not after), otherwise the
DISubprogram ends up with no DICompileUnit and LLVM fails to
compile.

This patch also enables DWARF versions above v2 for platforms that
do support these versions (i.e. everything but OS X and Android). It
also defines the "Debug Info Version" that matches the LLVM release.