Compiler: be more permissive with initialize as macro methods
Right now an initialize that is a macro method fails to type-check
because the compiler can't prove that instance vars inside it
were initialized (because the macro will be expanded later)

This change makes the compiler initially ignore those methods,
and check that instance vars were proplery initialized later on.

This allows, for example, generating an initialize method
from a JSON pull parser by traversing a type's instance vars,
something that can't currently be done.