Zlib: split into Adler32, CRC32, Flate, Gzip and Zlib types
Crystal provides access to the Adler32, CRC32, DEFLATE,
gzip and zlib algorithms/formats. It currently does
so by binding to zlib (also known as libz). However, zlib is
just an implementation detail: this shouldn't leak to type
and method names because if we eventually decide to
change the library used to implement these, or maybe
implement stuff in pure Crystal, we'll be stuck with this
name.

So, here we split the contents of the Zlib module into:
- Adler32, for the Adler32 checksum algorithm
- CRC32, for the CRC32 checksum algorithm
- Flate: for the DEFLATE compression format (RFC 1951),
providing Reader and Writer types (Inflate and Deflate could also
work, but Reader and Writer are more obvious and consistent.)
Flate is also the name used by Go to provide the same
functionality, so it will be familiar to some.
- Gzip: for the gzip archive format (RFC 1952), which is just
a small wrapper (header and checksum) around the DEFLATE
format. Reader and Writer are provided, together with access
to the first gzip header.
- Zlib: for the zlib archive format (RFC 1950), which is just
a small wrapper around the DEFLATE format too (here the
format is also named the same as the C library, which brings
a lot of confusion). Reader and Writer are provided.

By doing this we also remove the need to know how to use the
zlib C library, which requires users to provide a cryptic `windowBits`
argument to choose the desierd format.

Finally, we rename HTTP::DeflateHandler to HTTP::CompressHandler
because that's what it does: it compress responses in either
gzip or DEFLATE, but not always DEFLATE (so the name was
misleading).

All of this is a big breaking change, but should be easy to upgrade
existing code and makes the standard library more consistent and
organized.