String: fix inconsistency introduced by lchomp. Fixes #3729
Now we have:
- chomp: removes trailing '\n' or \'r\n'
- chomp(suffix): removes suffix (if suffix is '\n' it acts as `chomp()`)
- rchop: removes last char
- rchop(suffix): removes suffix
- lchop: removes first char
- lchop(prefix): removes prefix

We remove `'lchomp`

We also have:
- strip: removes leading and trailing whitespace
- strip(char): removes all leading and trailing occurrences of char
- strip(string): removes all leading and trailing occurences of any char of string
- lstrip, rstrip: similar to the above three, but only in one side of the string

With this we have pretty much covered all ways to remove things from the front/end of a String in a consistent way. There's `chomp` that has no right/left version, but that came from Ruby and is very specific to removing a trailing newline. The chomp(arg) variant is kept because one can do `io.gets(delimiter, chomp: true)` and that would be equivalent to doying `io.gets(delimiter).chomp(delimiter)`.