Separate platform-specific parts of IO::FileDescriptor
In order to add windows support to Crystal, IO::FileDescriptor cannot depend on
posix APIs. Here, we refactor out the platform-specific parts of
IO::FileDescriptor into a mixin module in Crystal::System. Some
platform-specific methods are refactored out of methods which contain both
platform-specific and cross-platform behaviour. These methods are prefixed with
`system_` and are private.